+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 -   Make QR Codes follow the theme colour
 -   Fix modal width on mobile
+-   Add QR Code image option
 
 # v2.4.0